move copy to copy folder, move hard coded copy to copy
